pyproject.toml
[project]
name = "autoclap"
version = "0.1.0"
description = "👏AI-powered clapperboard tracking in video"
authors = [
    {name = "Eumgill98",email = "wjdghcks02@gmail.com"}
]
license = {text = "GPL-3.0 license"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic (>=2.12.5,<3.0.0)",
    "ultralytics (>=8.3.249,<9.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "paddleocr (>=3.3.2,<4.0.0)",
    "paddlepaddle (>=3.3.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
